Move generate_block impl to block production
